requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.4.0
drf-yasg>=1.21.0
Pillow>=10.0.0
PyMySQL>=1.1.0
django-filter>=24.0
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.1.0
python-decouple>=3.8
